Clean: Merge STEX ConnectivityMap
