+++ src/dewarping/CMakeLists.txt
@@ -16,7 +16,6 @@ SET(
         STEX_AffineImageTransform.cpp STEX_AffineImageTransform.h
         STEX_AffineTransform.cpp STEX_AffineTransform.h
         STEX_AffineTransformedImage.cpp STEX_AffineTransformedImage.h
-        STEX_ConnectivityMap.cpp STEX_ConnectivityMap.h
         STEX_GaussBlur.cpp STEX_GaussBlur.h
         STEX_HomographicTransform.h
         STEX_PlugHoles.h
