+++ src/imageproc/ConnectivityMap.h
@@ -21,6 +21,7 @@
 
 #include "Connectivity.h"
 #include "FastQueue.h"
+#include "foundation/GridAccessor.h"
 #include <QSize>
 #include <QColor>
 #include <Qt>
@@ -151,6 +152,32 @@ public:
     {
         return m_pData ? &m_data[0] : 0;
     }
+
+    /**
+     * \brief Returns a reference to the label at position (x, y).
+     */
+    uint32_t& operator()(int x, int y)
+    {
+        return m_pData[m_stride * y + x];
+    }
+
+    /**
+     * \brief Returns a const reference to the label at position (x, y).
+     */
+    uint32_t const& operator()(int x, int y) const
+    {
+        return m_pData[m_stride * y + x];
+    }
+
+    /**
+     * \brief Provides integration with rasterOpGeneric().
+     */
+    GridAccessor<uint32_t const> accessor() const;
+
+    /**
+     * \brief Provides integration with rasterOpGeneric().
+     */
+    GridAccessor<uint32_t> accessor();
 
     /**
      * \brief Returns non-padded dimensions of the map.
